+++ src/lib/display/display.h
@@ -42,6 +42,23 @@
 #define FONT_ADDR_E 0xE0
 #define FONT_ADDR_F 0xF0
 
+#define CASE_FONT_ADDR_0 0x00          // Address after going through the bitshift used to seperate them from the directive flags.
+#define CASE_FONT_ADDR_1 0x01          // Included as preprocessor defines because that simplifies the actual runtime      
+#define CASE_FONT_ADDR_2 0x02
+#define CASE_FONT_ADDR_3 0x03
+#define CASE_FONT_ADDR_4 0x04
+#define CASE_FONT_ADDR_5 0x05
+#define CASE_FONT_ADDR_6 0x06
+#define CASE_FONT_ADDR_7 0x07
+#define CASE_FONT_ADDR_8 0x08
+#define CASE_FONT_ADDR_9 0x09
+#define CASE_FONT_ADDR_A 0x0A
+#define CASE_FONT_ADDR_B 0x0B
+#define CASE_FONT_ADDR_C 0x0C
+#define CASE_FONT_ADDR_D 0x0D
+#define CASE_FONT_ADDR_E 0x0E
+#define CASE_FONT_ADDR_F 0x0F
+
 #define DIRECTIVE_NORMAL 0x01           // Define symbols for font operation directives; these should be universal in all three print functions.
 #define DIRECTIVE_REVERSED 0x02
 #define DIRECTIVE_NEGATIVE 0x03